create component script refactor
